Fixed the code to get rid of the compiler warning messages

+++ Collectors/CopyingCollector.cpp
@@ -92,7 +92,7 @@ void CopyingCollector::enqueueAllRoots(){
 	vector<Object*> roots;
 	for(i = 0 ; i < NUM_THREADS ; i++){
 		roots = myObjectContainer->getRoots(i);
-		for(j = 0 ; j < roots.size(); j++){
+		for(j = 0 ; j < (int)roots.size(); j++){
 			currentObj = roots[j];
 			if(currentObj && !currentObj->getVisited()){
 				currentObj->setVisited(true);
